Population getters, s2 getter, rename of geometry output models
