+++ examples/exmp037_s2/inp.xyz
@@ -1,0 +1,5 @@
+3
+
+C     -0.08810    0.00000    0.00005
+H     -0.47256    0.17155    1.00984
+H     -0.47256    0.78879   -0.65351
